fix msvc 2005 level 3 warnings
darcs-hash:20060908213220-16063-e2341d9d1300d62e047c5647318f065127795215.gz

+++ libsofia-sip-ua/su/su_alloc.c
@@ -1126,7 +1126,7 @@ void *su_realloc(su_home_t *home, void *data, isize_t size)
   su_alloc_t *sua;
   su_block_t *sub;
   size_t p;
-  int term = -size;
+  int term = -((int)size);
 
   if (!home) 
     return realloc(data, size);
